added egg (ellipse) with collision with the dice

+++ src/load_resources.c
@@ -30,7 +30,7 @@ void load_resources()
     }
 
     has_sill_texture = true;
-    sill_texture_x = screen_width / 2 - sill_texture.width / 2;
-    sill_texture_y = screen_height / 2 - sill_texture.height / 2;
+    sill_texture_x = (int)screen_width / 2 - sill_texture.width / 2;
+    sill_texture_y = (int)screen_height / 2 - sill_texture.height / 2;
     say(__FILE__, func, __LINE__, LOG_INFO, "done");
 }
